keep fixing

+++ proj/src/controller/video/graphics.h
@@ -9,7 +9,7 @@ vbe_mode_info_t mode_info;
 int (set_graphic_mode)(uint16_t submode);
 int (set_frame_buffer)(uint16_t mode, uint8_t** frame_buffer);
 int (paint_pixel)(uint16_t x, uint16_t y, uint32_t color, uint8_t* frame_buffer);
-int (fill_rectangle)(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint32_t color, uint8_t* frame_buffer);
+int (fill_rectangle)(int32_t x, int32_t y, uint16_t width, uint16_t height, uint32_t color, uint8_t* frame_buffer);
 int (get_mode_info)(uint16_t mode, vbe_mode_info_t *mode_info);
 uint32_t (frame_buffer_index)(uint16_t x, uint16_t y);
 uint32_t (pixel_index)(uint16_t x, uint16_t y);
